מלי רביעי ערב

+++ Maly_and_Feigi/src/component/Todos.jsx
@@ -8,13 +8,14 @@ function Todos() {
     //לקצר את הפונקציה בסינאפ
     //לקצר מערך טודו קטן
     // IהI פייגי רוצה לסדר את עניני    
-    const userDetailes = useLocation();
+    const { state } = useLocation();
     const [userTodos, setUserTodos] = useState([]);
     const [showAdditionForm, setShowAdditionForm] = useState(false);
     // const [editables, setEditables] = useState([]);
     const [todos, setTodos] = useState([]);
     // const [stringSearch, setStringSearch] = useState();
     const [nextId, setNextId] = useState();
+    const [searchType, setSearchType] = useState();
 
     useEffect(() => {
         //fech next id
@@ -29,7 +30,7 @@ function Todos() {
             }).catch(ex => alert(ex))
 
         //fech todos
-        fetch(`http://localhost:3000/todos?userId=${userDetailes.state.id}`)
+        fetch(`http://localhost:3000/todos?userId=${state.id}`)
             .then(response => {
                 if (!response.ok)
                     throw 'Error' + response.status + ': ' + response.statusText;
@@ -123,15 +124,19 @@ function Todos() {
     // }
 
     function deleteTask(userIndex, i, id) {
-        fetch(`http://localhost:3000/todos/${id}`, {
-            method: 'DELETE'
-        }).then(response => {
-            if (!response.ok)
-                throw 'Error' + response.status + ': ' + response.statusText;
-        }).then(() => {
-            setUserTodos((prev) => [...prev.slice(0, userIndex), null, ...prev.slice(userIndex + 1, prev.length)])
-            setTodos((prev) => [...prev.slice(0, i), ...prev.slice(i + 1, prev.length)])
-        }).catch((ex) => alert(ex));
+        if (confirm('Are you sure you want to delete this todo from the database?')) {
+            fetch(`http://localhost:3000/todos/${id}`, {
+                method: 'DELETE'
+            }).then(response => {
+                if (!response.ok)
+                    throw 'Error' + response.status + ': ' + response.statusText;
+            }).then(() => {
+                setUserTodos((prev) => [...prev.slice(0, userIndex), null, ...prev.slice(userIndex + 1, prev.length)])
+                setTodos((prev) => [...prev.slice(0, i), ...prev.slice(i + 1, prev.length)])
+            }).catch((ex) => alert(ex));
+        } else {
+            return;
+        }
     }
 
     function changeEditable(i) {
@@ -167,7 +172,7 @@ function Todos() {
         fetch(`http://localhost:3000/todos/${id}`, {
             method: 'PATCH',
             body: JSON.stringify({
-                title:title.value,
+                title: title.value,
                 completed: completed.checked
             }),
             headers: {
@@ -184,39 +189,41 @@ function Todos() {
     }
 
     function searchTodos(event) {
-        event.preventDefault()
-        // const{searchById,searchByTitle,searchByComplited}=event.target;
-        // console.log(searchById)
-        // console.log(searchByTitle)
-        // console.log(searchByComplited)
-        //יום אחר לשלב בין החיפושים או לנקות את שאר האינפוטים
         let foundsArr;
+        let foundIndex;
+        console.log(event.target.value)
         switch (event.target.name) {
-            // case "all":
-            //     foundsArr = userTodos.map((t, i) => { if (t != null) return { ...t, i: i, editable: false } })
-            //     setTodos(foundsArr.filter(t => t != undefined));
-            //     break;
-            case "searchById":
-                // if(target.value)
-                // setTodos()
-                // foundIndex = userTodos.findIndex(t => t != null && t.id.includes(target.value))
-                // setTodos([{ ...userTodos[foundIndex], i: foundIndex, editable: false }])
-                foundsArr = userTodos.map((t, i) => { if (t != null && t.id.includes(event.target.value)) return { ...t, i: i, editable: false } })
-                setTodos(foundsArr.filter(t => t != null))
-                console.log(todos)
+            case "id":
+                foundIndex = userTodos.findIndex(t => t != null && t.id == event.target.value)
+                setTodos([{ ...userTodos[foundIndex], i: foundIndex, editable: false }])
                 break;
-            case "searchByTitle":
+            case "title":
                 foundsArr = userTodos.map((t, i) => { if (t != null && t.title.includes(event.target.value)) return { ...t, i: i, editable: false } })
                 setTodos(foundsArr.filter(t => t != null))
-                console.log(todos)
                 break;
-            case "searchByComplited":
-                foundsArr = userTodos.map((t, i) => { if (t != null && `${t.completed}` == event.target.value) return { ...t, i: i, editable: false } })
+            case "searchBycompleted":
+                foundsArr = userTodos.slice()
+                    .map((t, i) => { if (t != null && t.completed) return { ...t, i: i, editable: false } })
+                setTodos(foundsArr.filter(t => t != null))
+                break;
+            case "notComplete":
+                foundsArr = userTodos.map((t, i) => { if (t != null && !t.completed) return { ...t, i: i, editable: false } })
                 setTodos(foundsArr.filter(t => t != null))
                 break;
         }
     }
 
+    function search(event) {
+        let foundsArr;
+        if (event.target.value == "all") {
+            foundsArr = userTodos.map((t, i) => { if (t != null) return { ...t, i: i, editable: false } })
+            setTodos(foundsArr.filter(t => t != null));
+            setSearchType();
+        }
+        else
+            setSearchType(event.target.value);
+        console.log(event.target.value)
+    }
     return (<>
         <br /><br />
         <button onClick={() => (setShowAdditionForm(prev => !prev))}>Add task</button>
@@ -232,71 +239,75 @@ function Todos() {
         <br />
         <label htmlFor='sort' >order by</label>
         <select onChange={sortTodos} name="sort">
-            <option value="all"> </option>
+            <option value="all" > </option>
             <option value="id">id</option>
             <option value="alphabet">alphabet</option>
             <option value="completed">completed</option>
             <option value="random">random</option>
         </select>
 
-        {/* <label htmlFor='search'>search</label>
-        <input type="text" name="search" onChange={event => setStringSearch(event.target.value)} />
-        <label htmlFor='search' >by</label>
-        <select onChange={searchTodo} name="search">
-            <option value="all"></option>
+        <label htmlFor='search' >search by</label>
+        <select onChange={search} name="search">
+            <option value="all" onClick={searchTodos}></option>
             <option value="id">id</option>
             <option value="title">title</option>
             <option value="completed">completed</option>
-        </select> */}
-        <form onChange={searchTodos}>
-            <label htmlFor='searchById'>search by id</label>
-            {/* <input type="text" name="searchById" onChange={event => searchTodos(event.target)} /> */}
-            <input type="text" name="searchById" />
-
-
-            <label htmlFor='searchByTitle'>search by title</label>
-            {/* <input type="text" name="searchByTitle" onChange={event => searchTodos(event.target)} /> */}
-            <input type="text" name="searchByTitle" />
-
-
-            {/* <label htmlFor='searchByComplited'>search by complited</label>
-        <input type="text" name="searchByComplited" onChange={event => setStringSearch(event.target.value)} /> */}
-
-            <p>search by complited</p>
-            <input type="radio" name="searchByComplited" value="true" />
-            <label htmlFor="complited">complited</label>
-            <input type="radio" name="searchByComplited" value="false" />
-            <label htmlFor="notComplete">not complete</label>
-            {/* <input onChange={event => searchTodos(event.target)} type="radio" name="searchByComplited" value="all" />
-        <label htmlFor="all">all</label> */}
-
-            {/* <p>search by complited</p>
-            <input type="radio" onChange={event => searchTodos(event.target)} name="searchByComplited" value="true" />
-            <label htmlFor="complited">complited</label>
-            <input onChange={event => searchTodos(event.target)} type="radio" name="searchByComplited" value="false" />
-            <label htmlFor="notComplete">not complete</label> */}
-            {/* <input onChange={event => searchTodos(event.target)} type="radio" name="searchByComplited" value="all" />
-        <label htmlFor="all">all</label> */}
-            <br />
-        </form>
+        </select>
+        <br />
+        {searchType ? (searchType == "completed" ?
+            <>
+                <label htmlFor="completed">completed</label>
+                <input type="radio" name="searchBycompleted" value="true" onChange={event => searchTodos(event)} />
+                <label htmlFor="notCompleted">not completed</label>
+                <input type="radio" name="searchBycompleted" value="false" onChange={event => searchTodos(event)} />
+            </> : <input type="text" name={searchType} onChange={event => searchTodos(event)} />)
+            : <></>}
 
         {todos.length == 0 ? <h2>There are no tasks</h2>
             : todos.map((todo, i) => {
                 return (
                     <form key={i} onSubmit={(event) => updateTask(event, todo.i, i, todo.id)}>
                         <span style={{ marginRight: 10 }}>{todo.id}: </span>
-                        {todo.editable ?
+                        {todo.editable ? <>
                             <input name="title" type="text" defaultValue={todo.title} style={{ width: 300 }} />
-                            : <span>{todo.title} </span>}
-                        <input name="completed" type="checkbox" disabled={!todo.editable} defaultChecked={todo.completed} />
+                            <input name="completed" type="checkbox" defaultChecked={todo.completed} /></>
+                            : <><span>{todo.title} </span>
+                                <input name="completed" type="checkbox" disabled={true} checked={todo.completed} /></>}
                         <img src={edit} onClick={() => changeEditable(i)} />
                         <img onClick={() => deleteTask(todo.i, i, todo.id)} src={trash} />
                         {todo.editable && <button type="submit" >update</button>}
                         <br /><br />
                     </form>)
-            })
-        }</>)
-    // }
+            })}</>)
 }
 
 export default Todos;
+
+
+
+// {/* <form onSubmit={searchTodos}>
+// // <label htmlFor='searchById'>search by id</label>
+// <input type="text" name="searchById" onChange={event => searchTodos(event.target)} />
+// {/* <input type="text" name="searchById" /> */}
+// {/* <label htmlFor='searchByTitle'>search by title</label> */}
+// {/* <input type="text" name="searchByTitle" onChange={event => searchTodos(event.target)} /> */}
+// {/* <input type="text" name="searchByTitle" /> */}
+// {/* {/* <label htmlFor='searchBycompleted'>search by completed</label> */}
+// {/* <input type="text" name="searchBycompleted" onChange={event => setStringSearch(event.target.value)} /> */} */}
+// {/* <p>search by completed</p>
+// <input type="radio" name="searchBycompleted" value="true" />
+// <label htmlFor="completed">completed</label>
+// <input type="radio" name="searchBycompleted" value="false" />
+// <label htmlFor="notComplete">not complete</label> */}
+// {/* <input onChange={event => searchTodos(event.target)} type="radio" name="searchBycompleted" value="all" />
+// <label htmlFor="all">all</label> */}
+// {/* <button type="submit">search</button> */}
+// {/* <p>search by completed</p>
+// <input type="radio" onChange={event => searchTodos(event.target)} name="searchBycompleted" value="true" />
+// <label htmlFor="completed">completed</label>
+// <input onChange={event => searchTodos(event.target)} type="radio" name="searchBycompleted" value="false" />
+// <label htmlFor="notComplete">not complete</label> */}
+// {/* <input onChange={event => searchTodos(event.target)} type="radio" name="searchBycompleted" value="all" />
+// // <label htmlFor="all">all</label> */}
+// {/* <br /> */}
+// //     </form >
